requirements.txt
# Function dependencies, for example:
google-cloud-documentai==0.3.0
google-cloud-storage==1.35.0
